+++ src/test/java/bank/recommendationservice/fintech/repository/RecommendationsRepositoryTest.java
@@ -1,6 +1,5 @@
 package bank.recommendationservice.fintech.repository;
 
-import bank.recommendationservice.fintech.testrepository.RecommendationsRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
